backend/pyproject.toml
[tool.poetry]
name = "jarvis-backend"
version = "0.1.0"
description = "JARVIS FastAPI backend: OpenAI chat, task classification, desktop agent"
authors = []
readme = "README.md"
packages = [{ include = "*", from = "." }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.109.0"
uvicorn = { extras = ["standard"], version = ">=0.27.0" }
python-dotenv = ">=1.0.0"
openai = ">=1.12.0"
httpx = ">=0.26.0"
python-multipart = ">=0.0.9"
pyautogui = ">=0.9.54"
mss = ">=9.0.0"
pillow = ">=10.0.0"
websockets = ">=12.0"
playwright = ">=1.40.0"
langgraph = ">=0.2.0"
langchain-openai = ">=0.2.0"
langchain-core = ">=0.3.0"
chromadb = ">=0.4.0"
tiktoken = ">=0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
